requirements.txt
asgiref==3.9.2
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.12.4
Pillow==10.1.0